Private classes for HB and SH
