+++ test/test_intensify/test_successive_halving.py
@@ -8,13 +8,13 @@ import time
 from ConfigSpace import Configuration, ConfigurationSpace
 from ConfigSpace.hyperparameters import UniformIntegerHyperparameter
 
-from smac.scenario.scenario import Scenario
-from smac.tae.execute_func import ExecuteTAFuncDict
 from smac.intensification.abstract_racer import RunInfoIntent
-from smac.intensification.successive_halving import SuccessiveHalving
-from smac.runhistory.runhistory import RunHistory
-from smac.tae import StatusType
+from smac.intensification.successive_halving import SuccessiveHalving, _SuccessiveHalving
+from smac.runhistory.runhistory import RunHistory, RunInfo, RunValue
+from smac.scenario.scenario import Scenario
 from smac.stats.stats import Stats
+from smac.tae import StatusType
+from smac.tae.execute_func import ExecuteTAFuncDict
 from smac.utils.io.traj_logging import TrajLogger
 
 from .test_eval_utils import eval_challenger
@@ -31,7 +31,350 @@ def get_config_space():
     return cs
 
 
+def target_from_run_info(RunInfo):
+    value_from_config = sum([a for a in RunInfo.config.get_dictionary().values()])
+    return RunValue(
+        cost=value_from_config,
+        time=0.5,
+        status=StatusType.SUCCESS,
+        starttime=time.time(),
+        endtime=time.time() + 1,
+        additional_info={}
+    )
+
+
 class TestSuccessiveHalving(unittest.TestCase):
+
+    def setUp(self):
+        unittest.TestCase.setUp(self)
+
+        self.rh = RunHistory()
+        self.cs = get_config_space()
+        self.config1 = Configuration(self.cs,
+                                     values={'a': 7, 'b': 11})
+        self.config2 = Configuration(self.cs,
+                                     values={'a': 13, 'b': 17})
+        self.config3 = Configuration(self.cs,
+                                     values={'a': 0, 'b': 7})
+        self.config4 = Configuration(self.cs,
+                                     values={'a': 29, 'b': 31})
+
+        self.scen = Scenario({"cutoff_time": 2, 'cs': self.cs,
+                              "run_obj": 'runtime',
+                              "output_dir": ''})
+        self.stats = Stats(scenario=self.scen)
+        self.stats.start_timing()
+
+        # Create the base object
+        self.SH = SuccessiveHalving(
+            stats=self.stats,
+            traj_logger=TrajLogger(output_dir=None, stats=self.stats),
+            rng=np.random.RandomState(12345),
+            deterministic=False,
+            run_obj_time=False,
+            instances=[1, 2, 3, 4, 5],
+            n_seeds=2,
+            initial_budget=2,
+            max_budget=5,
+            eta=2,
+        )
+
+    def test_initialization(self):
+        """Makes sure that a proper _SH is created"""
+
+        # We initialize the SH with zero intensifier_instances
+        self.assertEqual(len(self.SH.intensifier_instances), 0)
+
+        # Add an instance to check the _SH initialization
+        self.assertTrue(self.SH._add_new_instance(num_workers=1))
+
+        # Parameters properly passed to _SH
+        self.assertEqual(len(self.SH.intensifier_instances[0].inst_seed_pairs), 10)
+        self.assertEqual(self.SH.intensifier_instances[0].initial_budget, 2)
+        self.assertEqual(self.SH.intensifier_instances[0].max_budget, 5)
+
+    def test_process_results_via_sourceid(self):
+        """Makes sure source id is honored when deciding
+        which _SH will consume the result/run_info"""
+        # Mock the _SH so we can make sure the correct item is passed
+        for i in range(10):
+            self.SH.intensifier_instances[i] = mock.Mock()
+
+        # randomly create run_infos and push into SH. Then we will make
+        # sure they got properly allocated
+        for i in np.random.choice(list(range(10)), 30):
+            run_info = RunInfo(
+                config=self.config1,
+                instance=0,
+                instance_specific="0",
+                cutoff=None,
+                seed=0,
+                capped=False,
+                budget=0.0,
+                source_id=i,
+            )
+
+            # make sure results aren't messed up via magic variable
+            # That is we check only the proper _SH has this
+            magic = time.time()
+
+            result = RunValue(
+                cost=1,
+                time=0.5,
+                status=StatusType.SUCCESS,
+                starttime=1,
+                endtime=2,
+                additional_info=magic
+            )
+            self.SH.process_results(
+                run_info=run_info,
+                incumbent=None,
+                run_history=self.rh,
+                time_bound=None,
+                result=result,
+                log_traj=False
+            )
+
+            # Check the call arguments of each sh instance and make sure
+            # it is the correct one
+
+            # First the expected one
+            self.assertEqual(
+                self.SH.intensifier_instances[i].process_results.call_args[1]['run_info'], run_info)
+            self.assertEqual(
+                self.SH.intensifier_instances[i].process_results.call_args[1]['result'], result)
+            all_other_run_infos, all_other_results = [], []
+            for j in range(len(self.SH.intensifier_instances)):
+                # Skip the expected _SH
+                if i == j:
+                    continue
+                if self.SH.intensifier_instances[j].process_results.call_args is None:
+                    all_other_run_infos.append(None)
+                else:
+                    all_other_run_infos.append(
+                        self.SH.intensifier_instances[j].process_results.call_args[1]['run_info'])
+                    all_other_results.append(
+                        self.SH.intensifier_instances[j].process_results.call_args[1]['result'])
+            self.assertNotIn(run_info, all_other_run_infos)
+            self.assertNotIn(result, all_other_results)
+
+    def test_get_next_run_single_SH(self):
+        """Makes sure that a single _SH returns a valid config"""
+
+        challengers = [self.config1, self.config2, self.config3, self.config4]
+        for i in range(30):
+            intent, run_info = self.SH.get_next_run(
+                challengers=challengers,
+                incumbent=None, chooser=None, run_history=self.rh,
+                num_workers=1,
+            )
+
+            # Regenerate challenger list
+            challengers = [c for c in challengers if c != run_info.config]
+
+            if intent == RunInfoIntent.WAIT:
+                break
+
+            # Add the config to self.rh in order to make SH aware that this
+            # config/instance was launched
+            self.rh.add(
+                config=run_info.config,
+                cost=10,
+                time=0.0,
+                status=StatusType.RUNNING,
+                instance_id=run_info.instance,
+                seed=run_info.seed,
+                budget=run_info.budget,
+            )
+
+        # We should not create more _SH intensifier_instances
+        self.assertEqual(len(self.SH.intensifier_instances), 1)
+
+        # We are running with:
+        # 'all_budgets': array([2.5, 5. ]) -> 2 intensifier_instances per config top
+        # 'n_configs_in_stage': [2.0, 1.0],
+        # This means we run int(2.5) + 2.0 = 4 runs before waiting
+        self.assertEqual(i, 4)
+
+    def test_get_next_run_dual_SH(self):
+        """Makes sure that two  _SH can properly coexist and tag
+        run_info properly"""
+
+        # Everything here will be tested with a single _SH
+        challengers = [self.config1, self.config2, self.config3, self.config4]
+        for i in range(30):
+            intent, run_info = self.SH.get_next_run(
+                challengers=challengers,
+                incumbent=None, chooser=None, run_history=self.rh,
+                num_workers=2,
+            )
+
+            # Regenerate challenger list
+            challengers = [c for c in challengers if c != run_info.config]
+
+            # Add the config to self.rh in order to make SH aware that this
+            # config/instance was launched
+            if intent == RunInfoIntent.WAIT:
+                break
+            self.rh.add(
+                config=run_info.config,
+                cost=10,
+                time=0.0,
+                status=StatusType.RUNNING,
+                instance_id=run_info.instance,
+                seed=run_info.seed,
+                budget=run_info.budget,
+            )
+
+        # We create a second sh intensifier_instances as after 4 runs, the _SH
+        # number zero needs to wait
+        self.assertEqual(len(self.SH.intensifier_instances), 2)
+
+        # We are running with:
+        # 'all_budgets': array([2.5, 5. ]) -> 2 intensifier_instances per config top
+        # 'n_configs_in_stage': [2.0, 1.0],
+        # This means we run int(2.5) + 2.0 = 4 runs before waiting
+        # But we have 2 successive halvers now!
+        self.assertEqual(i, 8)
+
+    def test_add_new_instance(self):
+        """Test whether we can add a _SH and when we should not"""
+
+        # By default we do not create a _SH
+        # test adding the first instance!
+        self.assertEqual(len(self.SH.intensifier_instances), 0)
+        self.assertTrue(self.SH._add_new_instance(num_workers=1))
+        self.assertEqual(len(self.SH.intensifier_instances), 1)
+        self.assertIsInstance(self.SH.intensifier_instances[0], _SuccessiveHalving)
+        # A second call should not add a new _SH
+        self.assertFalse(self.SH._add_new_instance(num_workers=1))
+
+        # We try with 2 _SH active
+
+        # We effectively return true because we added a new _SH
+        self.assertTrue(self.SH._add_new_instance(num_workers=2))
+
+        self.assertEqual(len(self.SH.intensifier_instances), 2)
+        self.assertIsInstance(self.SH.intensifier_instances[1], _SuccessiveHalving)
+
+        # Trying to add a third one should return false
+        self.assertFalse(self.SH._add_new_instance(num_workers=2))
+        self.assertEqual(len(self.SH.intensifier_instances), 2)
+
+    def _exhaust_run_and_get_incumbent(self, sh, rh, num_workers=2):
+        """
+        Runs all provided configs on all intensifier_instances and return the incumbent
+        as a nice side effect runhistory/stats are properly filled
+        """
+        challengers = [self.config1, self.config2, self.config3, self.config4]
+        incumbent = None
+        for i in range(100):
+            try:
+                intent, run_info = sh.get_next_run(
+                    challengers=challengers,
+                    incumbent=None, chooser=None, run_history=rh,
+                    num_workers=num_workers,
+                )
+            except ValueError as e:
+                # Get configurations until you run out of them
+                print(e)
+                break
+
+            # Regenerate challenger list
+            challengers = [c for c in challengers if c != run_info.config]
+
+            if intent == RunInfoIntent.WAIT:
+                break
+
+            result = target_from_run_info(run_info)
+            rh.add(
+                config=run_info.config,
+                cost=result.cost,
+                time=result.time,
+                status=result.status,
+                instance_id=run_info.instance,
+                seed=run_info.seed,
+                budget=run_info.budget,
+            )
+            incumbent, inc_perf = sh.process_results(
+                run_info=run_info,
+                incumbent=incumbent,
+                run_history=rh,
+                time_bound=100.0,
+                result=result,
+                log_traj=False,
+            )
+        return incumbent, inc_perf
+
+    def test_parallel_same_as_serial_SH(self):
+        """Makes sure we behave the same as a serial run at the end"""
+
+        # Get the run_history for a _SH run:
+        rh = RunHistory()
+        stats = Stats(scenario=self.scen)
+        stats.start_timing()
+        _SH = _SuccessiveHalving(
+            stats=stats,
+            traj_logger=TrajLogger(output_dir=None, stats=stats),
+            rng=np.random.RandomState(12345),
+            deterministic=False,
+            run_obj_time=False,
+            instances=[1, 2, 3, 4, 5],
+            n_seeds=2,
+            initial_budget=2,
+            max_budget=5,
+            eta=2,
+        )
+        incumbent, inc_perf = self._exhaust_run_and_get_incumbent(_SH, rh)
+
+        # Just to make sure nothing has changed from the _SH side to make
+        # this check invalid:
+        # We add config values, so config 3 with 0 and 7 should be the lesser cost
+        self.assertEqual(incumbent, self.config3)
+        self.assertEqual(inc_perf, 7.0)
+
+        # Do the same for SH, but have multiple _SH in there
+        # _add_new_instance returns true if it was able to add a new _SH
+        # We call this method twice because we want 2 workers
+        self.assertTrue(self.SH._add_new_instance(num_workers=2))
+        self.assertTrue(self.SH._add_new_instance(num_workers=2))
+        incumbent_psh, inc_perf_psh = self._exhaust_run_and_get_incumbent(self.SH, self.rh)
+        self.assertEqual(incumbent, incumbent_psh)
+
+        # This makes sure there is a single incumbent in SH
+        self.assertEqual(inc_perf, inc_perf_psh)
+
+        # We don't want to loose any configuration, and particularly
+        # we want to make sure the values of _SH to SH match
+        self.assertEqual(len(self.rh.data), len(rh.data))
+
+        # We are comparing exhausted single vs parallel successive
+        # halving runs. The number and type of configs should be the same
+        # and is enforced as a dictionary key argument check. The number
+        # of runs will be different ParallelSuccesiveHalving has 2 _SH intensifier_instances
+        # yet we make sure that after exhaustion, the budgets a config was run
+        # should match
+        configs_sh_rh = {}
+        for k, v in rh.data.items():
+            config_sh = rh.ids_config[k.config_id]
+            if config_sh not in configs_sh_rh:
+                configs_sh_rh[config_sh] = []
+            if v.cost not in configs_sh_rh[config_sh]:
+                configs_sh_rh[config_sh].append(v.cost)
+        configs_psh_rh = {}
+        for k, v in self.rh.data.items():
+            config_psh = self.rh.ids_config[k.config_id]
+            if config_psh not in configs_psh_rh:
+                configs_psh_rh[config_psh] = []
+            if v.cost not in configs_psh_rh[config_psh]:
+                configs_psh_rh[config_psh].append(v.cost)
+
+        # If this dictionaries are equal it means we have all configs
+        # and the values track the numbers and actual cost!
+        self.assertDictEqual(configs_sh_rh, configs_psh_rh)
+
+
+class Test_SuccessiveHalving(unittest.TestCase):
 
     def setUp(self):
         unittest.TestCase.setUp(self)
@@ -59,7 +402,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         """
             test parameter initializations for successive halving - instance as budget
         """
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats,
             traj_logger=TrajLogger(output_dir=None, stats=self.stats),
             rng=np.random.RandomState(12345), deterministic=False, run_obj_time=False,
@@ -77,7 +420,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         """
             test parameter initialiations for successive halving - real-valued budget
         """
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats,
             traj_logger=TrajLogger(output_dir=None, stats=self.stats),
             rng=np.random.RandomState(12345), deterministic=True, run_obj_time=False,
@@ -95,7 +438,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         """
             test parameter initialiations for successive halving - real-valued budget, high initial budget
         """
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats,
             traj_logger=TrajLogger(output_dir=None, stats=self.stats),
             rng=np.random.RandomState(12345), deterministic=True, run_obj_time=False,
@@ -117,7 +460,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         # runtime as budget (no param provided)
         with self.assertRaisesRegex(ValueError,
                                     "requires parameters initial_budget and max_budget for intensification!"):
-            SuccessiveHalving(
+            _SuccessiveHalving(
                 stats=self.stats,
                 traj_logger=TrajLogger(output_dir=None, stats=self.stats),
                 rng=np.random.RandomState(12345), deterministic=True, run_obj_time=False,
@@ -125,7 +468,7 @@ class TestSuccessiveHalving(unittest.TestCase):
 
         # eta < 1
         with self.assertRaisesRegex(ValueError, "eta must be greater than 1"):
-            SuccessiveHalving(
+            _SuccessiveHalving(
                 stats=self.stats,
                 traj_logger=TrajLogger(output_dir=None, stats=self.stats),
                 rng=np.random.RandomState(12345), deterministic=True, run_obj_time=False,
@@ -134,7 +477,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         # max budget > instance-seed pairs
         with self.assertRaisesRegex(ValueError,
                                     "Max budget cannot be greater than the number of instance-seed pairs"):
-            SuccessiveHalving(
+            _SuccessiveHalving(
                 stats=self.stats,
                 traj_logger=TrajLogger(output_dir=None, stats=self.stats),
                 rng=np.random.RandomState(12345), deterministic=True, run_obj_time=False,
@@ -144,7 +487,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         """
             test _top_k() for configs with same instance-seed-budget keys
         """
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345),
             instances=[1, 2], initial_budget=1)
@@ -182,7 +525,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         """
             test _top_k() for configs with different instance-seed-budget keys
         """
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345),
             instances=[1, 2], initial_budget=1)
@@ -201,7 +544,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         """
             test _top_k() for not enough configs to generate for the next budget
         """
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345),
             instances=[1], initial_budget=1)
@@ -220,7 +563,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         """
             test _top_k() for not enough configs to generate for the next budget
         """
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None, run_obj_time=False,
             rng=np.random.RandomState(12345), eta=2, num_initial_challengers=4,
             instances=[1], initial_budget=1, max_budget=10)
@@ -264,7 +607,7 @@ class TestSuccessiveHalving(unittest.TestCase):
 
         taf.runhistory = self.rh
 
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345), deterministic=True, run_obj_time=False,
             cutoff=1, instances=[1, 2], initial_budget=1, max_budget=2, eta=2)
@@ -315,7 +658,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         """
             test get_next_run for higher stages of SH iteration
         """
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345), deterministic=True, run_obj_time=False,
             cutoff=1, instances=[1], initial_budget=1, max_budget=2, eta=2)
@@ -339,7 +682,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         """
             test update_stage - initializations for all tracking variables
         """
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345), deterministic=True, run_obj_time=False,
             cutoff=1, instances=[1], initial_budget=1, max_budget=2, eta=2)
@@ -371,13 +714,13 @@ class TestSuccessiveHalving(unittest.TestCase):
         self.assertIsInstance(intensifier.configs_to_run, list)
         self.assertEqual(len(intensifier.configs_to_run), 0)
 
-    @unittest.mock.patch.object(SuccessiveHalving, '_top_k')
+    @unittest.mock.patch.object(_SuccessiveHalving, '_top_k')
     def test_update_stage_2(self, top_k_mock):
         """
             test update_stage - everything good is in state do not advance
         """
 
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345), deterministic=True, run_obj_time=False,
             cutoff=1, initial_budget=1, max_budget=4, eta=2, instances=None)
@@ -401,7 +744,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         self.assertEqual(len(intensifier.success_challengers), 0)
         self.assertEqual(len(intensifier.do_not_advance_challengers), 0)
 
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345), deterministic=True, run_obj_time=False,
             cutoff=1, initial_budget=1, max_budget=4, eta=2, instances=None)
@@ -438,7 +781,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         taf = ExecuteTAFuncDict(ta=target, stats=self.stats, run_obj='quality')
         taf.runhistory = self.rh
 
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats,
             traj_logger=TrajLogger(output_dir=None, stats=self.stats),
             rng=np.random.RandomState(12345), deterministic=True, run_obj_time=False,
@@ -489,7 +832,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         taf = ExecuteTAFuncDict(ta=target, stats=self.stats, run_obj="runtime")
         taf.runhistory = self.rh
 
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats,
             traj_logger=TrajLogger(output_dir=None, stats=self.stats),
             rng=np.random.RandomState(12345), deterministic=True, cutoff=1,
@@ -554,7 +897,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         self.assertEqual(self.stats.ta_runs, 3)
         self.assertEqual(list(self.rh.data.values())[2][2], StatusType.TIMEOUT)
 
-    @mock.patch.object(SuccessiveHalving, '_top_k')
+    @mock.patch.object(_SuccessiveHalving, '_top_k')
     def test_eval_challenger_capping(self, patch):
         """
             test eval_challenger with adaptive capping and all configurations capped/crashed
@@ -571,7 +914,7 @@ class TestSuccessiveHalving(unittest.TestCase):
                                 abort_on_first_run_crash=False)
         taf.runhistory = self.rh
 
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats,
             traj_logger=TrajLogger(output_dir=None, stats=self.stats),
             rng=np.random.RandomState(12345), deterministic=True, cutoff=1,
@@ -650,7 +993,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         taf = ExecuteTAFuncDict(ta=target, stats=self.stats, run_obj="runtime")
         taf.runhistory = self.rh
 
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats,
             traj_logger=TrajLogger(output_dir=None, stats=self.stats),
             rng=np.random.RandomState(12345), deterministic=False, cutoff=1,
@@ -759,7 +1102,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         taf = ExecuteTAFuncDict(ta=target, stats=self.stats, run_obj="quality")
         taf.runhistory = self.rh
 
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats,
             traj_logger=TrajLogger(output_dir=None, stats=self.stats),
             rng=np.random.RandomState(12345), run_obj_time=False,
@@ -858,7 +1201,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         """
             test _compare_config for default incumbent selection design (highest budget so far)
         """
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345), run_obj_time=False,
             instances=[1], initial_budget=1, max_budget=2, eta=2)
@@ -902,7 +1245,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         self.assertEqual(inc, self.config3)
 
         # Test that the state is only based on the runhistory
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345),
             instances=[1], initial_budget=1)
@@ -926,7 +1269,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         """
 
         # select best on any budget
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345), run_obj_time=False,
             instances=[1], initial_budget=1, max_budget=2, eta=2, incumbent_selection='any_budget')
@@ -949,7 +1292,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         self.assertEqual(self.config1, inc)
 
         # select best on highest budget only
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345), run_obj_time=False,
             instances=[1], initial_budget=1, max_budget=4, eta=2, incumbent_selection='highest_budget')
@@ -983,7 +1326,7 @@ class TestSuccessiveHalving(unittest.TestCase):
 
         taf.runhistory = self.rh
         # select best on any budget
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345), run_obj_time=False,
             instances=list(range(20)), initial_budget=3, max_budget=10, eta=2)
@@ -1073,7 +1416,7 @@ class TestSuccessiveHalving(unittest.TestCase):
 
         taf.runhistory = self.rh
         # select best on any budget
-        intensifier = SuccessiveHalving(
+        intensifier = _SuccessiveHalving(
             stats=self.stats, traj_logger=None,
             rng=np.random.RandomState(12345), run_obj_time=False,
             instances=list(range(10)), initial_budget=2, max_budget=10, eta=2)
@@ -1123,3 +1466,7 @@ class TestSuccessiveHalving(unittest.TestCase):
         # not registered any, so for sure we have to wait
         # For all runs to be completed before moving to the next stage
         self.assertEqual(intent, RunInfoIntent.WAIT)
+
+
+if __name__ == "__main__":
+    unittest.main()
